requirements.txt
Django==5.0.6
djangorestframework==3.15.1
dj-rest-auth==6.0.0
django-allauth==0.63.3
django-extensions==3.2.3
django-cors-headers==4.3.1
setuptools==70.0.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-redis==5.4.0
python-decouple==3.8
psycopg2-binary==2.9.9

requests
ipython
